Add files via upload
RSVP-Movies
SQL queries performed on IMDb database to provide recommendations to RSVP Movies based on insights.

Problem Introduction
RSVP Movies is an Indian film production company which has produced many super-hit movies. They have usually released movies for the Indian audience but for their next project, they are planning to release a movie for the global audience in 2022.

The production company wants to plan their every move analytically based on data and have approached you for help with this new project. You have been provided with the data of the movies that have been released in the past three years. You must analyze the data set and draw meaningful insights that can help them start their new project.

You are a data analyst and an SQL expert. You must use SQL to analyze the given data and give recommendations to RSVP Movies based on the insights. For your convenience, the entire analytics process has been divided into four segments, where each segment leads to significant insights from different combinations of tables. The questions in each segment with business objectives are written in the script.

Data Set and Database Creation
Download the IMDb dataset.
The first tab contains the ERD and the table details. Study that carefully and understand the relationships between the table.
Inspect each table given in the subsequent tabs and understand the features associated with each of them.
Open your MySQL Workbench and start writing the DDL and DML commands to create the database.
If you don't wish to perform the data loading part, you can directly download the SQL script file containing all the commands and data required for the database creation and start directly with the querying.